requirements.txt
python-dotenv==1.0.0
SpeechRecognition==3.10.0
google-generativeai==0.3.2
pyttsx3==2.90
twilio==8.3.0
geocoder==1.38.1
flask==3.0.2
keyboard==0.13.5
